Add second iteration of related links AB test
This commit adds the second iteration of the related links A/B test, as described at https://www.gov.uk/help/cookies#multivariate-testing. The cookie age is set to 12 hours initially, with all traffic to the `A` variant, to help put less stress on the system. We will increase the A/B split afterwards.