Add third iteration of Related Links test
This commit adds the third iteration of the related links test as an A/B/C, described at https://www.gov.uk/help/cookies#multivariate-testing. The cookie age is set to 12 hours initially, with all traffic to the `B` variant, to help put less stress on the system. We will increase the B/C split afterwards.

The A variant will not be actively set by the CDN, but is instead being used to measure whether any users "drop through" somehow in our A/B testing framework, as we have previously seen users with an A variant cookie access B variant links.