Restore ability to purge URLs by Fastly web UI
After a discussion with Fastly Support we learnt that we don't have the
ability to purge individual URLs from Fastly via the web UI because of
our VCL configuration. We actually return a 403 as the requests are not
in the allowed list.

Fastly support suggested that we change our return of error 403 to
instead set a head a header that authentication is required. Without a
valid token to authenticate these will deny the request and return a
message that authentication is needed.