Cache non-GET/HEAD requests with 404/405 responses
Trello: https://trello.com/c/6P4c8RiN/396

In response to a P1 issue in February 2019, this
change will mean that non-GET/HEAD requests that
result in a 404 Not Found or 405 Method Not Allowed
status will be cached in the same way as GET/HEAD
requests are currently.

The change to vcl_recv means that requests with any
method (excluding FASTLYPURGE) will be looked up
in the cache, rather than being passed straight
through to origin to be handled by an application.

Alone, this would mean that e.g. POST requests
would be cached, and future POST requests would
be served a response from the cache, rather than
making it to an application. That would be chaotic.

The change to vcl_fetch ensures that we only cache
non-GET/HEAD requests if they return a 404 or 405.

This means that if someone POSTs to /education
in an interesting way, rather than asking the
application to handle this and return a 405
(or potentially a 404), the cache will handle this
and these requests wont use up application resources.

The change to vcl_hash ensures that all requests
that use a different request method (POST/GET/DELETE)
are cached uniquely, so a GET request won't return
the cached response to a POST request.

There are risks to this change.
It is potentially quite dangerous to cache non-GET/HEAD requests.

If backends respond with incorrect status codes
we will cache POST requests. We should hope that
backends will return correct responses, but this
cache logic will be a side effect of using HTTP
status codes incorrectly that developers may not
anticipate.

Example: A 404 might wrongly be returned from successful requests.
If an endpoint returns a 404 after a delete request
rather than say e.g. 200, 202, 410, or some other code
one could choose, then we would cache this response.
If the content is recreated following a deletion,
then this cache logic may prevent a further deletion
of the same object.

There are lots of instances of this kind of thing,
please let me know of other cases where this might
lead to more severe side effects.

Another risk is that private data returned in 404
responses might be exposed. For example, if a 404
response also includes a payload submitted as part
of a POST request body.

It is worth noting that publishing applications are not
served through fastly.

Another risk is that a higher number of lookup() calls on
the varnish cache will have an adverse effect. Now all
requests will go through a lookup, I don't know whether
this will impact performance of the cache.