Remove VCL code for cacheing non-GET/HEAD requests
This reverts commit 689fe7eb4c800ac9fb4393c150133d1c3285389a.

We found a bug in this code that prevents FASTLYPURGE requests
from working as expected. In the vcl_hash we include the request
method in the hash key we use to store and lookup objects in the cache.
This works fine for GET/HEAD/POST requests because we can cache
the responses to these requests as separate objects (a post response
might be different to a get response).

However, it means that FASTLYPURGE requests lookup items in the cache
using FASTLYPURGE as part of the hash, preventing purges from working
as expected.

Rather than attempting to patch this by adding more complexity to the
VCL, we've decided to approach the problem we were trying to solve in
a different way.

Thanks to @kevindew for working on this with me.

https://trello.com/c/7jFNfgr6/1346