Replace Jenkinsfile with github action
This project can be switched to github actions as it has no downstream
deployments in accordance with RFC 123 [1].

A casualty of this change is that this project will no longer receive
release tags. We may no use of these tags however and their numbering is
coupled to Jenkins so the only problem is that we end up with a history
in GitHub that will just fizzle out.

[1]: https://github.com/alphagov/govuk-rfcs/blob/master/rfc-123-github-actions-ci.md