Put the user in an A/B bucket based on whether they're logged in
We want to show an accounts-related header on /transition and in the
transition checker, with either sign in or sign out links.

A/B testing feels like a slightly weird way to do this, because it's
not an A/B test, but it is an established technique we have for
displaying variants of GOV.UK pages which plays nicely with caching.

We should revisit this when we address the larger topic of how to be
logged in across GOV.UK (rather than logged in to one app).