Reset proxy headers to avoid spoofing alerts
Previously a user could manually specify these headers, which could
result in a spoofing error for Rails apps trying to detect the client
IP address [1]. Although the use of the headers by the end user is
valid, we can't trust and shouldn't care about any proxies that the
request has been through before reaching our network. This ensures
the proxy headers are reset before the request continues.

Trying this manually (before/after) shows that:

- Fastly does not have a default value for the Client-IP header, so
it's consistent to enforce that it is unset.

- Fastly pre-populates X-Forwarded-For with the "actual" client IP,
so we should preserve that (but ignore any pre-existing header).

Note that HTTP headers are not case sensitive.

[1]: https://github.com/alphagov/govuk_app_config/pull/171