Add CookielessAATest
This commit adds `CookielessAATest`, which is using [the recommended approach from Fastly ]() for running A/B tests without cookies, instead relying on an MD5 hash of a client IP address and user-agent info. Due to the fact that this test is to be run only when a user has not yet accepted cookies, the cookie expiry for this test is set to `0` and the configuration is defined separately within the `multivariate_tests` partial.