Remove postcode formatting
This removes the postcode formatting for requests to
/find-coronavirus-local-restrictions in favour of a JavaScript one done
within the collections app: https://github.com/alphagov/collections/pull/2156

The switch to a JavaScript approach was chosen so that the code can be
kept with the app (with tests) and to improve user experience (this CDN one will
produce confusing outcomes for invalid postcodes). The trade-off for the
JS one is that is less aggressive in it's formatting and, by nature of
being client side, not something that is forced upon a user. The
expectation is that these trade offs won't produce any significant
variance in traffic.