Recognise 307 responses as cacheable on www.gov.uk
GOV.UK responds with 307 for some responses to indicate a temporary
redirect. Fastly doesn't currently support them but suggested this
change.

I decided not to include the similar 303 response in here as this seemed
a bit more contentious given it is expected to originate from a POST
request: https://github.com/varnishcache/varnish-cache/issues/3340 and I
don't believe we have a need.

This same change could be applied to our other VCL files but I'm not
aware of them having a need currently so it seems superfluous to
duplicate this for each.