Add test spec for Test environment
We are creating a new ECS-based platform for GOV.UK.

This PR adds the test spec for the new fastly service for this new platform.

We intend only for the www vcl to be used for now for the Test environment even though
the other vcls are being generated by the test task.

Ref:
[1] [trello card](https://trello.com/c/TjJOKSKp/404-basic-auth-for-integration-environment)