Allow UI compression settings to persist on deploy
Fastly allows compressions settings to be toggled in the UI. We'd like
to investigate enabling this on staging, and if it looks good roll it
out to production.

The Default compression policy says:

> Compress text to speed the transfer of data. By default we compress:
css, js, html, eot, ico, otf, ttf, json, and svg formats.

> We support Brotli (the default) and gzip compression formats. The
setting depends upon browser support.

Toggling this and looking at the config diff shows that it puts
everything in the `gzip` section of the config (a bit badly named, since
it supports Brotli as well as GZip, but I imagine that's historical).

Currently our deployment script deletes any `gzip` objects it finds in
the UI, so we can't use the feature. Removing this will allow us to
snowflake the config in.

In the long run we should use terraform for this instead.