Only set GOVUK-Account-Session if cookie is present
Currently we're setting this header on every request, which means it's
usually getting set to the empty string.  This has resulted in
finder-frontend making a request to account-api every time the brexit
checker results page gets loaded, because it just checks if the header
is set at all, not if it's set with a nonempty value.