Add custom GOVUK-Account-Session-Exists request header
We will use this to show the appropriate navigation (signed in vs
signed out), and `Vary` based on it so we only cache one copy for all
signed in users and one copy for all signed out users, rather than one
copy each for all signed in users.

I don't think we need to add a case for:

    if (resp.http.Vary ~ "GOVUK-Account-Session-Exists")

because `~` is a regex match, so the `~ "GOVUK-Account-Session"` check
should be sufficient.

See also the progressive enhancement ADR for account-api[1].

[1] https://github.com/alphagov/account-api/blob/main/docs/adr/002-progressive-enhancement.md