Bypass cache if there is a flash message to show
The account session cookie is set by govuk_personalisation, and is in
one of two formats:

- "#{base64(encrypt(session))}"
- "#{base64(encrypt(session))$$#{flash}"

We want to bypass the cache if there is a flash message, so that the
app can render it.  We considered doing flash messages as a form of
progressive enhancement, but decided that doing them server-side made
more sense and was simpler.