Use a custom request header for account flash caching purposes
Previously, we bypassed Fastly when there is a flash message in the
session.  But this isn't enough: there's also our Varnish.

A request header with a corresponding `Vary` response header is the
standard way to handle a request that is conditionally cacheable so,
rather than implementing a further special case for our Varnish, just
go for the standard.

Then, the way a page which shows a flash and account navigation but
doesn't use a user's attributes will work is:

1. It reads and decodes the account session header
2. It sets `Vary: GOVUK-Account-Session-Exists`
3. It sets `Vary: GOVUK-Account-Session-Flash`
4. It fetches the appropriate template from static
5. It reads the flash to determine what will be shown

This is better than just setting `Vary: GOVUK-Account-Session` (which
would also work), because there's no need to cache one
page-with-empty-flash for each individual logged in user.

This replaces #356.