Allow setting the Host header for requests to the EKS backend.
The Ingress-managed ALB which fronts traffic to the Kubernetes cluster
serves 404 when the hostname doesn't match what it expects. It's simpler
for now to set the hostname in Fastly than to switch to using
`defaultBackend` in the Ingress spec.

Off by default so we don't affect any of the existing configs.