Define the allowed_ip_addresses ACL in all non-prod environments.
This ACL is referred to in all the non-production configurations. Since
we're adding configurations for integration-eks and staging-eks, it's
cleaner and safer to specify the environments where we don't define the
ACL than to keep adding to a list of ones where we do. Defining an
unused ACL is harmless anyway.