Allow specifying CIDRs in the allowed_ip_addresses ACL.
This is needed in order to add the new egress addresses for the DR VPN;
otherwise we'd have to add 136 individual IP addresses and it'd be
infeasible to maintain.

VCL has a funky syntax for CIDR netblocks, hence the ugly hack in the
template. Normally I'd try to find a neater way (e.g. define a function
outside the template) but we're hopefully about to throw this code away.
Open to better suggestions.