Enable Cookie stripping in Fastly rather than Varnish
This is needed to enable the stripping from most applications except a
select few.  This is a feature that previously existed in varnish but is
not being moved to Fastly to support the EKS replatforming of the apps.

Another pr was used to test this in the integration environment, and we
were able to see using tcpdump and wireshark that the stripping was
occuring at the appropriate level in the stack.

Original PR: https://github.com/alphagov/govuk-cdn-config/pull/373

Co-authored-by: smford <stephen.ford@digital.cabinet-office.gov.uk>
Co-authored-by: samsimpson1 <sam.simpson@digital.cabinet-office.gov.uk>
Co-authored-by: jibolaolu <seun.adiomowo@digital.cabinet-office.gov.uk>