Set X-Forwarded-Host headers
This ensures that Fastly sets the X-Forwarded-Host headers to a single
value that is the same original request host. This helps mitigate
against clients for spoofing the X-Forwarded-Host header and prevents
cache poisoning. This must be set in the vcl_miss and vcl_pass
subroutines as they occur before backend requests and have access to the
the `bereq` token. This assumes that backend requests are only for
client requests that have valid and non-spoofed hosts, which is fine as
Fastly will only be dealing with request that have to correct Host
header for the service.