Change RenderTemplate interface to accept local vars
This refactors usage of the RenderTemplate class to make it more
flexible so that it can be used by other services.

It has been changed from acting as a helper function to being a class
instance, this is so that we can extend it further to add in helper
methods on RenderTemplate that can be used (I intend to add a method to
render partials later in this PR).