Add #render_partial to RenderTemplate
This method is to allow templates being rendered to call
`render_partial` to render a partial template.

This aims to be somewhat similar in syntax in Rails, but is very simple
and thus probably won't handle much scrutiny.

As we aim for VCL files to be readable I've taken the rather unusual
step of adding an indentation argument that can be used to indent the
output of a template without a dev having to manually do anything. This
may prove to be a foolish idea if it causes problems with any multi-line
bits of VCL, but I think it'll be ok.