Consistently set headers at Fastly boundary for all services
This applies the same header setting approach for requests at the Fastly
boundary of all our Fastly services that proxy.

The motivation for this is to be clearer in what we trust header wise
and not risk situations where our applications can behave differently
based on headers a user sets (cache poisoning).

This moves the same client-IP header code that was in most of the files
to one place. This sets a header of True-Client-IP based on the IP that
connected to Fastly. We also remove the Client-IP header, I'm continuing
to do that based on the precedence but I couldn't identify what that
precedent was.

We set the X-Forwarded-For header to be the IP that communicated with
Fastly, prior to a request reaching Fastly we can't trust that the data
in here isn't a user specifying a fake IP address.

We reset the X-Forwarded-Host to be the host that Fastly received a
request for, this prevents a user pretending that the request originated
from a different host. By default Fastly will append to this header, as
Fastly doesn't know if the request to it is from a trusted source or
not.

We set X-Forwarded-Server to the Fastly server hostname. I'm not sure if
this is actually of value, but it seems better to override it than ever
be caught out by a user setting this and we think it's from within our
network.