Set proxy headers with original request headers
This fixes an issue where the X-Forwarded-Host header was incorrect, as
it was being set from the request Host header which was being re-written
previously. This moves the logic to set the proxy headers earlier in the
vcl_recv subroutine as to ensure the are using the original header
information.