Stop setting http_host for requests to assets-origin
This change depends on a corresponding Puppet PR [1]

This removes the setting of a HTTP Host header for when Fastly is
proxying to origin. This was previously needed because there wasn't
vhosts configured for the integration and staging asset hosts, so
proxying to them would fail.

It's not necessarily bad that we change the host here, my motivation is
more based on consistency as this is behaviour we don't follow in our
other CDN configs (www and bouncer). I believe this inconsistency
contributed to a recent incident [2] and thus believe that resolving it
we reduce the risk of further surprises caused by quirky configuration.

[1]: https://github.com/alphagov/govuk-puppet/pull/11871
[2]: https://docs.google.com/document/d/18kIgL2F0PPCZSjSCLH5z0hOq3Gn8wmCwxrlR1bOPric/edit#heading=h.e85n7w4jfehw