Drop query params for /alerts URLs
There is no requirement for query params to reach origin on the /alerts paths
and we don't require them for cache busting or other features within Fastly.

By dropping these, we improve our cache hit rate at the edge and our ability
to cater for large amounts of traffic.

This regex should match:

/alerts
/alerts/
/alerts/something
/alerts/something/longer

but not

/random
/alerts-go-missing
/alerts-go-missing/example

We're matching on the path, so we shouldn't need to worry about params here.