Configure AA test for ElasticSearch 6.7
The purpose of the A/A test is to test the % of allocation of traffic/sample population based on the current version of elastic (6.7) against itself. A (ES 6.7) vs A (ES6.7). The A/A is solely to identify how the approach allocates traffic. This will provide benchmark data to aid analysis in future AA tests.

Rendering app PR: https://github.com/alphagov/finder-frontend/pull/3053