Remove hard-coded bank holidays test code
The setting of a cookie for AB tests is not working correctly in the
dynanmic code. Hard-coding the test setup allowed us to determine that
the fault was somewhere in the code.

The code will be updated in the next commit to allow AB test cookies to
be set, so this test code can be removed.