Use flag to determine if a user has opted in to cookies
What stopped working?

The dynamic code used to generate cookies for AB tests first checks
that the user has consented to cookies. This value is stored in a
cookie called `cookie_policy`. If the `cookie_policy` cookie doesn't
contain a value, the code won't execute. That's what was happening
here, and had the effect of the AB test cookies not being created.

The Example AB test doesn't check the value of the `cookie_policy`
cookie, so that cookie was being set as expected.

Why did it stop working?

When a request is made to Fastly for a page, the header contains the
cookie information. However, if there's a cache-miss and we need to
call the application on origin (as there is when a cookie needs to be
set), we need to make sure that we don't send the cookies to Origin.

We do this to prevent the risk of applications caching user information
from the cookie and accidently sharing it with other users. Instead the
application reads the test variant that the user is in from the request
header. This header value is added by Fastly when it receives the request.

In the old world, pre-replatforming, there was another layer that sat
between Fastly and Origin; Varnish. As well as being another caching
layer, Varnish also did the work to strip the cookie information before
sending the request on to the application.

As the cookie stripping was happening in the Varnish layer, Fastly
always had access to the cookie information. That means that when the
request came back from Origin via Varnish, Fastly was still able to look
up the `cookie_policy` value and set the AB Test cookies.

The Varnish layer was removed with replatforming. This meant that the
cookies had to be stripped elsewhere before the request was sent to
Origin. As Fastly talks to Origin directly now, the cookie stripping
was added to Fastly [before being sent] to Origin. This initial test of
the cookie stripping occurred in integration and it was found that it
stopped AB Tests from working as it ran before the code to set the
AB test variant in the header (called from the multivariant
partial).

So the cookie stripping was moved to [after the header] value was
added. As the code now seemed to set the header correctly, it was
applied to all environments.

However the way VCL is structured and when each bit of the code is
executed means that setting cookies for AB tests stopped working.

Even though all of the generated VCL code is created in the same place,
it actually runs at the different times in the [life of a request].

The code to strip the cookies is in the `vcl_recv` function and is run
when the request is received.

The code that sets AB test cookies is in the `vcl_deliver` function and
is run against the response from Origin before delivering that response
to the user. By the time this code runs the cookie information from the
request is long gone.

Ideally, the cookie information pertaining to cookie consent in the
`cookie_policy` cookie, and any existing AB test cookies needs to be
preserved in Fastly so that the "deliver" code can access it.

How does adding a flag fix the problem?

It doesn't fix it, not completely.

A variable has been created to persist the `cookie_policy`. Having this
variable means that the AB Test cookies can be created.

However, this hasn't completely fixed the problem. As the rest of the
cookie information has been lost, it means that any code in the
`vcl_deliver` function that checks the value of `req.http.Cookie` will
always return `(null)`. That means that this check will always be
true:

```
req.http.Cookie !~ "ABTest-<%= test %>"
```

The consequence of this is that everytime the user visits a page in an
AB test, a new cookie will always be set for them. This renders the
cookie expiry time as useless, and means that it's likely that the user
could see or interact with multiple variants of a page. This is not
ideal in terms of user journey, or for testing user behaviour.

What are the next steps?

We need to find a way to stop generating an AB test cookie every time a
user visits the page, most likely by preserving the AB test cookie data
in `req.http.Cookie` before it's purged, but without having to hardcode
the AB test name everytime a new one is added.

[before being sent]: https://github.com/alphagov/govuk-cdn-config/commit/4330a2530bbd6878cfba6c4093f9e2e847633feb
[after the header]: https://github.com/alphagov/govuk-cdn-config/commit/ad64a8a4afdd3d951a3fc67101ad4d5b050b0f6b
[life of a request]: https://developer.fastly.com/learning/vcl/using/

Co-authored-by: Richard Towers <richard.towers@digital.cabinet-office.gov.uk>