Avoid using req.http.Cookie in vcl_deliver
As discussed in previous commits, the value of req.http.Cookie is unset
at the end of vcl_recv, so we can't use its value in vcl_deliver.

To avoid resending the Set-Cookie header for the ABTest cookie on every
response, we need to know if the Cookie was sent in the request. We're
adding an HTTP header to the request to track whether this cookie was
sent, and then checking this header in vcl_deliver instead of checking
req.http.Cookie.