Move always the same conditionals outside of loop
The check that we're opted in to cookies and that we're not the crawler
worker is the same regardless of which AB test we're running, so we can
move this out of the `ab_tests.each` block and only stamp the if
statement out once, even if there are mulitiple tests.

This shouldn't change the logic of the tests, it's just a refactoring.