Remove Licensify TLS hack
In March 2018, in 26a0edb0c1138b5cbedd741e396d942cf117cd8c,
code was added to intercept requests for URLs beginning with
"/apply-for-a-licence/", which:

> Added a request header containing the TLS version number for requests to the Licensify app
> (This is so we can block requests with TLS versions below 1.2 from submitting, for payment security reasons)

However, in December 2022, in
https://github.com/alphagov/govuk-puppet/pull/11931/commits/46d98aaec4eaf664f8a078ea12d2ec03ada4b18c,
we dropped TLS 1 and 1.1 support at the nginx level. So the
special case added just for Licensify is no longer necessary.

This was uncovered as part of the following Trello card:
https://trello.com/c/TbvosEgo/3101-staging-test-define-list-of-features-that-wont-currently-work-on-cloudfront