Update the % for ElasticSearch 6.7 AA test
The AA test is currently deployed at a 50/50 variant split.
This PR updates the AA test to throttle it the split at 0.5%.
The C variant has been added to make the difference between
the CDN traffic allocation and anything that falls into the Z
variant (handled by the rendering app) more explicit.